+++ rules-tests/Renaming/Rector/FunctionLike/RenameFunctionLikeParamWithinCallLikeArgRector/Fixture/skip_missing_parameters.php.inc
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rector\Tests\Renaming\Rector\FunctionLike\RenameFunctionLikeParamWithinCallLikeArgRector\Fixture;
+
+use SomeNamespace\SomeClass;
+
+SomeClass::someOtherCall('a', function () {
+});
+
+?>
