+++ rules-tests/Renaming/Rector/FunctionLike/RenameFunctionLikeParamWithinCallLikeArgRector/Fixture/skip_non_matching_class.php.inc
@@ -1,0 +1,9 @@
+<?php
+
+namespace Rector\Tests\Renaming\Rector\FunctionLike\RenameFunctionLikeParamWithinCallLikeArgRector\Fixture;
+
+SomeOtherClass::someCall(function ($qry) {
+    return $qry;
+});
+
+?>
