+++ rules-tests/Renaming/Rector/FunctionLike/RenameFunctionLikeParamWithinCallLikeArgRector/Fixture/skip_non_matching_method.php.inc
@@ -1,0 +1,11 @@
+<?php
+
+namespace Rector\Tests\Renaming\Rector\FunctionLike\RenameFunctionLikeParamWithinCallLikeArgRector\Fixture;
+
+use SomeNamespace\SomeClass;
+
+SomeClass::someOtherCall(function ($qry) {
+    return $qry;
+});
+
+?>
