--- rules-tests/Renaming/Rector/FunctionLike/RenameFunctionLikeParamWithinCallLikeParamRector/Fixture/skip_missing_parameters.php.inc
@@ -1,10 +1,0 @@
-<?php
-
-namespace Rector\Tests\Renaming\Rector\FunctionLike\RenameFunctionLikeParamWithinCallLikeParamRector\Fixture;
-
-use SomeNamespace\SomeClass;
-
-SomeClass::someOtherCall('a', function () {
-});
-
-?>
