--- rules-tests/Renaming/Rector/FunctionLike/RenameFunctionLikeParamWithinCallLikeParamRector/Fixture/skip_non_matching_class.php.inc
@@ -1,9 +1,0 @@
-<?php
-
-namespace Rector\Tests\Renaming\Rector\FunctionLike\RenameFunctionLikeParamWithinCallLikeParamRector\Fixture;
-
-SomeOtherClass::someCall(function ($qry) {
-    return $qry;
-});
-
-?>
