--- rules-tests/Renaming/Rector/FunctionLike/RenameFunctionLikeParamWithinCallLikeParamRector/Fixture/skip_non_matching_method.php.inc
@@ -1,11 +1,0 @@
-<?php
-
-namespace Rector\Tests\Renaming\Rector\FunctionLike\RenameFunctionLikeParamWithinCallLikeParamRector\Fixture;
-
-use SomeNamespace\SomeClass;
-
-SomeClass::someOtherCall(function ($qry) {
-    return $qry;
-});
-
-?>
